django/acts-security-system/requirements.txt
Django==2.2.4
docxtpl==0.6.3
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
pip==19.3.1
python-dateutil==2.8.1
python-docx==0.8.7
pytz==2019.3
setuptools==41.6.0
six==1.13.0
sqlparse==0.3.0
wheel==0.33.6
